Rename MuonAngleGenerator to MuonGenerator
