+++ source/generators/MuonGenerator.cc
@@ -1,8 +1,7 @@
 // ----------------------------------------------------------------------------
 // nexus | MuonGenerator.cc
 //
-// This class is the primary generator of muons following the angular
-// distribution at sea level. Angles are saved to be plotted later.
+// This class is the primary generator of muons
 //
 // The NEXT Collaboration
 // ----------------------------------------------------------------------------
@@ -15,6 +14,7 @@
 #include "FactoryBase.h"
 #include "RandomUtils.h"
 
+#include <G4Event.hh>
 #include <G4GenericMessenger.hh>
 #include <G4ParticleDefinition.hh>
 #include <G4RunManager.hh>
@@ -22,16 +22,18 @@
 #include <G4PrimaryVertex.hh>
 #include <G4Event.hh>
 #include <G4RandomDirection.hh>
+#include <Randomize.hh>
 
+#include "CLHEP/Units/SystemOfUnits.h"
 
 using namespace nexus;
 
 REGISTER_CLASS(MuonGenerator, G4VPrimaryGenerator)
 
-
 MuonGenerator::MuonGenerator():
   G4VPrimaryGenerator(), msg_(0), particle_definition_(0),
-  energy_min_(0.), energy_max_(0.), geom_(0), momentum_{}
+  use_lsc_dist_(true), rPhi_(NULL), energy_min_(0.),
+  energy_max_(0.), dist_name_("za"), user_dir_{}, bInitialize_(false), geom_(0), geom_solid_(0)
 {
   msg_ = new G4GenericMessenger(this, "/Generator/MuonGenerator/",
 				"Control commands of muongenerator.");
@@ -51,12 +53,76 @@ MuonGenerator::MuonGenerator():
   msg_->DeclareProperty("region", region_,
 			"Set the region of the geometry where the vertex will be generated.");
 
-  msg_->DeclarePropertyWithUnit("momentum", "mm",  momentum_,
-    "Set particle 3-momentum.");
+  msg_->DeclareProperty("use_lsc_dist", use_lsc_dist_,
+			"Distribute muon directions according to file?");
 
+  msg_->DeclareProperty("angle_file", ang_file_,
+			"Name of the file containing angular distribution.");
+  msg_->DeclareProperty("angle_dist", dist_name_,
+			"Name of the angular distribution histogram.");
+
+  msg_->DeclareProperty("user_dir",  user_dir_, "Set fixed muon direction.");
+
+  G4GenericMessenger::Command& rotation =
+    msg_->DeclareProperty("azimuth_rotation", axis_rotation_,
+			  "Angle between north and nexus z in anticlockwise");
+  rotation.SetUnitCategory("Angle");
+  rotation.SetParameterName("azimuth", false);
+  rotation.SetRange("azimuth>0.");
 
   DetectorConstruction* detconst = (DetectorConstruction*) G4RunManager::GetRunManager()->GetUserDetectorConstruction();
   geom_ = detconst->GetGeometry();
+
+}
+
+
+MuonGenerator::~MuonGenerator()
+{
+  delete msg_;
+}
+
+void MuonGenerator::LoadMuonDistribution()
+{
+  
+
+  // Check the input filename for the keyword Energy
+  size_t found = ang_file_.find("Energy");
+  if (found!=std::string::npos){
+
+    // Only angle option, but energy specified
+    if (dist_name_ == "za")
+      G4Exception("[MuonGenerator]", "LoadMuonDistribution()",
+                FatalException, " Angular + Energy file specified with angle_dist=za option selected, use angle_dist=zae ");
+  }
+  // File name does not contain the word Energy
+  else {
+    if (dist_name_ == "zae")
+      G4Exception("[MuonGenerator]", "LoadMuonDistribution()",
+                FatalException, " Angular file specified with angle_dist=zae option selected, use angle_dist=za ");
+  }
+
+  // Load in the data from csv file depending on 2D histogram sampling or 3D
+  if (dist_name_ == "za")
+    CSV_Reader_->LoadHistData2D(ang_file_, flux_, azimuths_, zeniths_, azimuth_smear_, zenith_smear_);
+
+  if (dist_name_ == "zae"){
+    CSV_Reader_->LoadHistData3D(ang_file_, flux_, azimuths_, zeniths_, energies_, azimuth_smear_, zenith_smear_, energy_smear_);
+    
+    // Check if the energy is in the desired range permitted by the binning range in the data file
+    CSV_Reader_->CheckVarBounds(ang_file_, energy_min_/GeV, energy_max_/GeV, "energy"); 
+
+  }
+
+  // Convert flux vector to arr
+  auto arr_flux = flux_.data();
+
+  // Initialise the Random Number Generator based on the flux distribution (in bin index)
+  fRandomGeneral_ = new G4RandGeneral( arr_flux, flux_.size() );
+
+}
+
+void MuonGenerator::InitMuonZenithDist()
+{
 
   // Create a vector of values finely spaced from 0 -> pi/2
   // Then take cos(x)*cos(x) to make a dist to sample from
@@ -67,98 +133,200 @@ MuonGenerator::MuonGenerator():
   }
 
   // Convert vector to arr
-  G4double arr_ang[v_angles.size()];
-  std::copy(v_angles.begin(), v_angles.end(), arr_ang);
+  auto arr_ang = v_angles.data();
 
   // Initialise the Random Number Generator based on cos(x)*cos(x) distribution
   fRandomGeneral_ = new G4RandGeneral( arr_ang, v_angles.size() );
 
 }
 
-
-MuonGenerator::~MuonGenerator()
+void MuonGenerator::SetupAngles()
 {
-
-  delete msg_;
+  // Rotation from the axes used in file.
+  // Rotates anticlockwise about Y.
+  rPhi_ = new G4RotationMatrix();
+  rPhi_->rotateY(-axis_rotation_);
 }
+
 
 void MuonGenerator::GeneratePrimaryVertex(G4Event* event)
 {
-  particle_definition_ = G4ParticleTable::GetParticleTable()->FindParticle(MuonCharge());
+
+  // Initalise RNG seeds and load file on the first event
+  if (!bInitialize_){
+
+    // Load in the Muon angular distribution from file
+    if (use_lsc_dist_){
+      std::cout << "[MuonGenerator]: Generating muons using lsc distribution loaded from file" << std::endl;
+      LoadMuonDistribution();
+    }
+    
+    // Initalise a cos^2 distribution to sample the zenith
+    if (!use_lsc_dist_ && (user_dir_ == G4ThreeVector{})){
+      std::cout << "[MuonGenerator]: Generating muons with uniform azimuth and cos^2 distribution for zenith " << std::endl;
+      InitMuonZenithDist();
+    }
+
+    // User specified muon direction
+    if (!use_lsc_dist_ && (user_dir_ != G4ThreeVector{})){
+      std::cout << "[MuonGenerator]: Generating muons with user specified direction " << std::endl;
+    }
+
+    // Set Initialisation
+    bInitialize_ = true;
+
+  }
+
+  // Init the rotation matrix
+  if (rPhi_ == NULL)
+    SetupAngles();
+
+  particle_definition_ =
+    G4ParticleTable::GetParticleTable()->FindParticle(MuonCharge());
+
   if (!particle_definition_)
     G4Exception("[MuonGenerator]", "SetParticleDefinition()",
                 FatalException, " can not create a muon ");
 
-  // Generate an initial position for the particle using the geometry
-  G4ThreeVector position = geom_->GenerateVertex(region_);
-  // Particle generated at start-of-event
-  G4double time = 0.;
-  // Create a new vertex
-  G4PrimaryVertex* vertex = new G4PrimaryVertex(position, time);
   // Generate uniform random energy in [E_min, E_max]
   G4double kinetic_energy = UniformRandomInRange(energy_max_, energy_min_);
 
-  // Particle propierties
-  G4double mass   = particle_definition_->GetPDGMass();
-  G4double energy = kinetic_energy + mass;
-  G4double pmod = std::sqrt(energy*energy - mass*mass);
+  // Particle properties
+  G4double mass          = particle_definition_->GetPDGMass();
+  G4double energy        = kinetic_energy + mass;
+  G4ThreeVector position = geom_->GenerateVertex(region_);
 
-  // Generate momentum direction in spherical coordinates
-  G4double theta = GetTheta();
-  G4double phi   = GetPhi();
+  // Set default momentum and angular variables
+  G4ThreeVector p_dir;
+  G4double zenith;
+  G4double azimuth;
 
-  // NEXT axis convention (z<->y) and generate with -y! towards the detector
-  G4double x = sin(theta) * cos(phi);
-  G4double y = -cos(theta);
-  G4double z = sin(theta) * sin(phi);
+  // Momentum, zenith, azimuth (and energy) from angular distribution file
+  if (use_lsc_dist_){
+    GetDirection(p_dir, zenith, azimuth, energy, kinetic_energy, mass);
+    position = geom_->GenerateVertex(region_);
+  }
+  else {
 
-  G4ThreeVector p_dir(x,y,z);
+    // User specified muon direction in some fixed direction
+    if ( user_dir_ != G4ThreeVector{}) {
+      p_dir   = user_dir_.unit();
+      zenith  = p_dir.getTheta();
+      azimuth = p_dir.getPhi() + pi; // change azimuth interval to be between 0, twopi
+    }
 
-  bool fixed_momentum = momentum_ != G4ThreeVector{};
-  // If user provides a momentum direction, this one is used
-  if (fixed_momentum) {
-    p_dir = momentum_.unit();
-    phi   = p_dir.getPhi() + pi; // change phi interval to be between 0, twopi
-    theta = p_dir.getTheta();
+    // Sample direction via cos^2 distribution for zenith, uniform azimuth
+    else {
+      zenith  = GetZenith();
+      azimuth = GetAzimuth(); // Returns from 0 to 2pi
+
+      // Calculate the vector components of the muon
+      p_dir.setX(sin(zenith) * sin(azimuth));
+      p_dir.setY(-cos(zenith));
+      p_dir.setZ(-sin(zenith) * cos(azimuth));
+
+      // Rotate about the Y-Axis
+      p_dir *= *rPhi_;
+
+    }
+
   }
 
-  G4ThreeVector p = pmod * p_dir;
+  G4double pmod   = std::sqrt(energy*energy - mass*mass);
+  G4double px = pmod * p_dir.x();
+  G4double py = pmod * p_dir.y();
+  G4double pz = pmod * p_dir.z();
+
+  // Particle generated at start-of-event
+  G4double time = 0.;
+  
+  // Create a new vertex
+  G4PrimaryVertex* vertex = new G4PrimaryVertex(position, time);
 
   // Create the new primary particle and set it some properties
   G4PrimaryParticle* particle =
-    new G4PrimaryParticle(particle_definition_, p.x(), p.y(), p.z());
+    new G4PrimaryParticle(particle_definition_, px, py, pz);
 
   // Add info to PrimaryVertex to be accessed from EventAction type class to make histos of variables generated here.
-  AddUserInfoToPV *info = new AddUserInfoToPV(theta, phi);
+  AddUserInfoToPV *info = new AddUserInfoToPV(zenith, azimuth);
 
   vertex->SetUserInformation(info);
 
   // Add particle to the vertex and this to the event
   vertex->SetPrimary(particle);
   event->AddPrimaryVertex(vertex);
-
 }
 
 
 G4String MuonGenerator::MuonCharge() const
 {
-  G4double rndCh = 2.3 *G4UniformRand(); //From PDG cosmic muons  mu+/mu- = 1.3
-  if (rndCh <1.3)
+
+  // Ratio of Mu+/Mu- is energy dependent, ranges from 1.3 to 1.5:
+  // https://arxiv.org/pdf/1111.6675.pdf
+  // Assume (approx) flat ratio up to energy range of interest ~6 TeV 
+  // mu+/mu- ~1.3
+
+  // Sample random number to give 1.3 to 1 ratio of Mu+/Mu-
+  G4double rndCh = 2.3 * G4UniformRand(); 
+  if (rndCh < 1.3)
     return "mu+";
   else
     return "mu-";
 }
 
 
-G4double MuonGenerator::GetTheta() const
+void MuonGenerator::GetDirection(G4ThreeVector& dir, G4double& zenith, G4double& azimuth,
+                      G4double& energy, G4double& kinetic_energy, G4double mass)
 {
-  G4double theta;
-  theta = fRandomGeneral_->fire()*pi/2;
-  return theta;
+
+  // Bool to check if zenith has a valid value. If not then resample
+  G4bool invalid_evt = true;
+
+  while(invalid_evt){
+
+    // Generate random index weighted by the bin contents
+    // Scale by flux vec size, then round to nearest integer to get an index
+    G4int RN_indx = Dist_Sampler_->GetRandBinIndex(fRandomGeneral_, flux_);
+
+    // Correct sampled values by Gaussian smearing
+    azimuth = Dist_Sampler_->Sample(azimuths_[RN_indx], true, azimuth_smear_[RN_indx]);
+    zenith  = Dist_Sampler_->Sample(zeniths_[RN_indx],  true, zenith_smear_[RN_indx]);
+
+    // Sample and update the energy if angle + energy option specified
+    if (dist_name_ == "zae"){
+      energy = Dist_Sampler_->Sample(energies_[RN_indx]*GeV, true, energy_smear_[RN_indx]*GeV);
+      kinetic_energy = energy - mass;
+
+    }
+
+    // Catch negative value and resample if so
+    if (zenith < 0.0){
+        invalid_evt = true;
+        continue;
+    }
+    else
+        invalid_evt = false;
+
+    // Calculate the vector components of the muon
+    dir.setX(sin(zenith) * sin(azimuth));
+    dir.setY(-cos(zenith));
+    dir.setZ(-sin(zenith) * cos(azimuth));
+
+    // Rotate about the Y-Axis
+    dir *= *rPhi_;
+
+  }
+
 }
 
 
-G4double MuonGenerator::GetPhi() const
+G4double MuonGenerator::GetZenith() const
+{
+  return fRandomGeneral_->fire()*pi/2;
+}
+
+
+G4double MuonGenerator::GetAzimuth() const
 {
   return twopi*G4UniformRand();
 }
